tests/wpt/tests/navigation-api/updateCurrentEntry-method/not-fully-active.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" src="/common/blank.html"></iframe>

<script>
async_test(t => {
  window.onload = t.step_func_done(() => {
    const wNavigation = iframe.contentWindow.navigation;
    const wDOMException = iframe.contentWindow.DOMException;

    iframe.remove();

    assert_throws_dom("InvalidStateError", wDOMException, () => {
      wNavigation.updateCurrentEntry({ state: 1 });
    });
    assert_equals(navigation.currentEntry.getState(), undefined);
  });
}, "updateCurrentEntry() must throw if the document is not fully active");
</script>
