LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/not-fully-active-expected.txt


PASS updateCurrentEntry() must throw if the document is not fully active
